--- a/pkg/streamingpromql/benchmarks/benchmarks.go
+++ b/pkg/streamingpromql/benchmarks/benchmarks.go
@@ -173,9 +173,9 @@ func TestCases(metricSizes []int) []BenchCase {
 		//{
 		//	Expr: "a_X or b_X{l=~'.*[0-4]$'}",
 		//},
-		//{
-		//	Expr: "a_X unless b_X{l=~'.*[0-4]$'}",
-		//},
+		{
+			Expr: "a_X unless b_X{l=~'.*[0-4]$'}",
+		},
 		{
 			Expr: "a_X and b_X{l='notfound'}",
 		},