--- a/pkg/streamingpromql/testdata/upstream/operators.test
+++ b/pkg/streamingpromql/testdata/upstream/operators.test
@@ -210,18 +210,15 @@ eval instant at 50m (http_requests{group="canary"} + 1) and ignoring(group, job)
 #	vector_matching_a{l="x"} 10
 #	vector_matching_a{l="y"} 20
 
-# Unsupported by streaming engine.
-# eval instant at 50m http_requests{group="canary"} unless http_requests{instance="0"}
-#	http_requests{group="canary", instance="1", job="api-server"} 400
-#	http_requests{group="canary", instance="1", job="app-server"} 800
+eval instant at 50m http_requests{group="canary"} unless http_requests{instance="0"}
+	http_requests{group="canary", instance="1", job="api-server"} 400
+	http_requests{group="canary", instance="1", job="app-server"} 800
 
-# Unsupported by streaming engine.
-# eval instant at 50m http_requests{group="canary"} unless on(job) http_requests{instance="0"}
+eval instant at 50m http_requests{group="canary"} unless on(job) http_requests{instance="0"}
 
-# Unsupported by streaming engine.
-# eval instant at 50m http_requests{group="canary"} unless on(job, instance) http_requests{instance="0"}
-#	http_requests{group="canary", instance="1", job="api-server"} 400
-#	http_requests{group="canary", instance="1", job="app-server"} 800
+eval instant at 50m http_requests{group="canary"} unless on(job, instance) http_requests{instance="0"}
+	http_requests{group="canary", instance="1", job="api-server"} 400
+	http_requests{group="canary", instance="1", job="app-server"} 800
 
 eval instant at 50m http_requests{group="canary"} / on(instance,job) http_requests{group="production"}
 	{instance="0", job="api-server"} 3
@@ -229,13 +226,11 @@ eval instant at 50m http_requests{group="canary"} / on(instance,job) http_reques
 	{instance="1", job="api-server"} 2
 	{instance="1", job="app-server"} 1.3333333333333333
 
-# Unsupported by streaming engine.
-# eval instant at 50m http_requests{group="canary"} unless ignoring(group, instance) http_requests{instance="0"}
+eval instant at 50m http_requests{group="canary"} unless ignoring(group, instance) http_requests{instance="0"}
 
-# Unsupported by streaming engine.
-# eval instant at 50m http_requests{group="canary"} unless ignoring(group) http_requests{instance="0"}
-# 	http_requests{group="canary", instance="1", job="api-server"} 400
-# 	http_requests{group="canary", instance="1", job="app-server"} 800
+eval instant at 50m http_requests{group="canary"} unless ignoring(group) http_requests{instance="0"}
+	http_requests{group="canary", instance="1", job="api-server"} 400
+	http_requests{group="canary", instance="1", job="app-server"} 800
 
 eval instant at 50m http_requests{group="canary"} / ignoring(group) http_requests{group="production"}
 	{instance="0", job="api-server"} 3